Add support for X-Return-Header command
X-Return-Header will force the specified header to be sent back on the response.
For instance, "X-Return-Header: Content-Type: application/json" will force
"Content-Type: application/json" to be returned on the response.

If more than one X-Return-Header is used with the same response header, those
headers are collated and return in turn.

For instance,
  X-Return-Header: Content-Type: application/json
  X-Return-Header: Content-Type: text/xml

will return a response with
  Content-Type: application/json
  Content-Type: text/xml