Add the ability to generate random-but-structured JSON
This gives the ability to test a client's response when
a server returns a large number of responses, either by design
or due to a bug on the server side.

However, since many client libraries bind data structures
to the structure of the JSON, this also includes a small language
for generating JSON that conforms to some structure but uses
randomized data.